refactor(stores): split dataStore and chatStore along their seams; one built-in tool table
chatStore.svelte.ts (3.4k lines) held three things. The pure half — the
checkpoint graph, message-pair derivation, tool timelines, summarization
markers, context formatting — is now stores/chatTimeline.ts (1.6k lines,
no reactive state). ChatSession and SessionRegistry stay in
chatStore.svelte.ts. Consumers import the derivation types from the new
module.

dataStore.svelte.ts (2.7k lines) loses its non-store residents:
- stores/agentDefaults.ts — DEFAULT_AGENT_ID, default agent factories,
  load-time agent normalization and name de-duplication
- stores/dataMigrations.ts — schema version + the migration table
- stores/staleGuidance.ts — the prompt-surface staleness detector

Built-in tool metadata lived twice, with different wording: the UI table in
agent/builtInToolMeta.ts and the stored defaults in dataStore. Both now
derive from one table in agent/tools/builtInToolDefaults.ts (display name,
UI summary, stored ToolConfig per tool); DEFAULT_TOOLS_CONFIG and the
read_content / search_notes description variants live there too.
