+++ src/agent/tools/manageNotes.ts
@@ -2,7 +2,8 @@ import { type App, normalizePath, type TFile } from "obsidian";
 import { tool } from "@langchain/core/tools";
 import type { RunnableConfig } from "@langchain/core/runnables";
 import { z } from "zod";
-import { DEFAULT_TOOLS_CONFIG, getData } from "../../stores/dataStore.svelte";
+import { getData } from "../../stores/dataStore.svelte";
+import { DEFAULT_TOOLS_CONFIG } from "./builtInToolDefaults";
 import { getPendingChangesStore } from "../../stores/pendingChangesStore.svelte";
 import type { PendingChange } from "../../types/shared";
 import { resolveVaultFileDetailed } from "../../utils/pathResolution";
