+++ src/agent/tools/searchNotes.ts
@@ -5,7 +5,8 @@ import type { SearchAlgorithm } from "../../types/plugin";
 import { getLexicalSearchService, waitForLexicalSearch } from "../../search/LexicalSearchService";
 import { rankSearchResults } from "../../search/finalSearchRanking";
 import { buildRecentBoostMap, getRecentNotes } from "../../search/recentNotes";
-import { getData, getSearchNotesDescription } from "../../stores/dataStore.svelte";
+import { getData } from "../../stores/dataStore.svelte";
+import { getSearchNotesDescription } from "./builtInToolDefaults";
 import { getPendingChangesStore } from "../../stores/pendingChangesStore.svelte";
 import { normalizeVaultPath } from "../../utils/pathUtils";
 import { getVectorStoreService, type SearchFilter, type SearchResult, waitForVectorStore } from "../../vectorstore";
